feat(sandbox): move the preview options into a side panel
The navbar keeps what plays: platform, media, and source. The skin, its
styling, where it comes from, and the compare axis move into the preview
header. Width, appearance, playback, captions, language, and the detected
preferences move into an Options panel beside the preview, toggled from
the navbar and remembered across loads, so nothing covers the player.
The width control now reaches 1360px and takes an exact value.
