+++ apps/sandbox/README.md
@@ -10,7 +10,7 @@ pnpm dev:sandbox             # sandbox + workspace package watch
 pnpm dev                     # also runs the docs site
 ```
 
-Open the printed URL. The root route renders an interactive shell — a navbar with dropdowns for platform (HTML, React, CDN), media (`video`, `hlsjs-video`, `audio`, etc.), skin, styling (CSS or Tailwind), and source — that previews the selected combination in an iframe. Every selection is in the URL, so a link reproduces a preview. Use the **Open** button to pop the preview out into its own tab.
+Open the printed URL. The root route renders an interactive shell — a navbar with dropdowns for platform (HTML, React, CDN), media (`video`, `hlsjs-video`, `audio`, etc.), and source, with the skin, its styling (CSS or Tailwind), where it comes from, and what to compare it against in the row above the preview — that previews the selected combination in an iframe. Every selection is in the URL, so a link reproduces a preview. Use the **Open** button to pop the preview out into its own tab.
 
 `app/media.ts` describes each media once: its label, player, element, which sources the picker offers, and which controls apply. The shell derives its constraints from that table, and the CDN page picks its bundles by the same id. Older links that say `?preset=` still resolve; `preset` now means the player preset a skin is built for (`video`, `audio`, `live-video`, `live-audio`).
 
@@ -18,13 +18,13 @@ Open the printed URL. The root route renders an interactive shell — a navbar w
 
 **Compare** renders two previews that differ on one axis and share everything else: CSS against Tailwind, one skin source against the next that can load the styling, Default against Minimal, or HTML against React. Each panel is its own iframe, so two stylesheets never share a document, and its header names the value it takes. The panels sit side by side once the preview is wide enough and stack below that; the toggle above them forces either. **Mirror playback**, a checkbox beside the layout toggle, carries play, pause, seeks, volume, mute, playback rate, and caption selection from the panel you touch to the other one; it moves state through the media element rather than pointer positions, so it works across skins and platforms. The URL carries `compare`, `layout`, and `mirror`.
 
-**Report** copies a markdown summary for a bug report and shows it in a dialog: the URL, the branch and commit the sandbox was served from, the selection in words, each panel's URL when comparing, browser, viewport, the detected preferences, and the last errors the preview frames relayed (uncaught errors, unhandled rejections, and `console.error`). The same preference badges sit at the bottom of **Player settings** and follow DevTools' rendering emulation live.
+**Report** copies a markdown summary for a bug report and shows it in a dialog: the URL, the branch and commit the sandbox was served from, the selection in words, each panel's URL when comparing, browser, viewport, the detected preferences, and the last errors the preview frames relayed (uncaught errors, unhandled rejections, and `console.error`). The same preference badges sit at the bottom of the **Options** panel and follow DevTools' rendering emulation live.
 
-**Width** sits in the navbar and sizes the player through `--sandbox-player-width`; until it is touched, a preview opens at its skin's own width. **Color scheme** and **Direction** are in **Player settings** and pin the preview's `color-scheme` and `dir`, where the defaults follow the operating system and the locale. All three travel in the URL (`width`, `scheme`, `dir`), so a direct page honours them too.
+**Width** sits at the top of the **Options** panel (the sliders icon in the navbar opens and closes it, and the choice is remembered), as a slider up to 1360px with a field for an exact value, and sizes the player through `--sandbox-player-width`; until it is touched, a preview opens at its skin's own width. **Color scheme** and **Direction** are in the same panel and pin the preview's `color-scheme` and `dir`, where the defaults follow the operating system and the locale. All three travel in the URL (`width`, `scheme`, `dir`), so a direct page honours them too.
 
-**Captions**, also in **Player settings**, adds one or two subtitle tracks to a video so the captions menu has something to show; the tracks are the page's, so no template spells them out. The source picker also carries a **Missing file** entry that opens the player's error dialog without waiting on a network.
+**Captions**, under **Playback** in the **Options** panel, adds one or two subtitle tracks to a video so the captions menu has something to show; the tracks are the page's, so no template spells them out. The source picker also carries a **Missing file** entry that opens the player's error dialog without waiting on a network.
 
-**Language** is in **Player settings** (gear icon) for every media (HTML, React, and CDN). **CDN** registers copy through `@videojs/html/cdn/i18n` (the same registry as the CDN player bundle), not source `@videojs/html/i18n`. After pulling template changes, restart `pnpm dev:sandbox` so `scripts/setup.ts` refreshes `src/` from `templates/`.
+**Language** is in the **Options** panel for every media (HTML, React, and CDN). **CDN** registers copy through `@videojs/html/cdn/i18n` (the same registry as the CDN player bundle), not source `@videojs/html/i18n`. After pulling template changes, restart `pnpm dev:sandbox` so `scripts/setup.ts` refreshes `src/` from `templates/`.
 
 The shell covers the main combinatorial matrix. One-off templates not in that matrix (e.g. `firefox-mse-repro`, `spf-segment-loading`, `hls-video-html`) are reachable by navigating directly to `/<template-name>/`. See `apps/sandbox/templates/` for the full list.
